add json wrapper for golang

+++ framework/cei_golang/src/impl/utils/stringutils.go
@@ -27,7 +27,7 @@ func ToString(value interface{}) string {
 	case uint32:
 		return strconv.FormatUint(uint64(v), 10)
 	case uint64:
-		return strconv.FormatUint(uint64(v), 10)
+		return strconv.FormatUint(v, 10)
 	case uint:
 		return strconv.FormatUint(uint64(v), 10)
 	case int8:
